fixed up the player camera quite a bit
