--- Assets/scripts/PlayerCamera2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bc123a0f7a2e76a4f98ccb830f8aea77
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
